requirements.txt
streamlit
pandas
requests
sentence-transformers
scikit-learn
torchvision
fpdf2
matplotlib
